Skip event and snapshot deletion if up to sequence number is zero

+++ core/src/main/scala/lerna/akka/entityreplication/raft/RaftActor.scala
@@ -987,23 +987,26 @@ private[raft] class RaftActor(
     if (settings.deleteOldEvents) {
       val deleteEventsToSequenceNr =
         math.max(0, success.metadata.sequenceNr - deleteBeforeRelativeSequenceNr)
-      if (log.isDebugEnabled) {
-        log.debug("[{}] Deleting events up to sequenceNr [{}]", currentState, deleteEventsToSequenceNr)
+      if (deleteEventsToSequenceNr > 0) {
+        if (log.isDebugEnabled) {
+          log.debug("[{}] Deleting events up to sequenceNr [{}]", currentState, deleteEventsToSequenceNr)
+        }
+        deleteMessages(deleteEventsToSequenceNr)
       }
-      deleteMessages(deleteEventsToSequenceNr)
     }
     if (settings.deleteOldSnapshots) {
-      val deletionCriteria = {
-        // Since this actor will use the snapshot with sequence number `metadata.sequenceNr` for recovery, it can delete
-        // snapshots with sequence numbers less than `metadata.sequenceNr`.
-        val deleteSnapshotsToSequenceNr =
-          math.max(0, success.metadata.sequenceNr - 1 - deleteBeforeRelativeSequenceNr)
-        SnapshotSelectionCriteria(minSequenceNr = 0, maxSequenceNr = deleteSnapshotsToSequenceNr)
+      // Since this actor will use the snapshot with sequence number `metadata.sequenceNr` for recovery, it can delete
+      // snapshots with sequence numbers less than `metadata.sequenceNr`.
+      val deleteSnapshotsToSequenceNr =
+        math.max(0, success.metadata.sequenceNr - 1 - deleteBeforeRelativeSequenceNr)
+      if (deleteSnapshotsToSequenceNr > 0) {
+        val deletionCriteria =
+          SnapshotSelectionCriteria(minSequenceNr = 0, maxSequenceNr = deleteSnapshotsToSequenceNr)
+        if (log.isDebugEnabled) {
+          log.debug("[{}] Deleting snapshots matching criteria [{}]", currentState, deletionCriteria)
+        }
+        deleteSnapshots(deletionCriteria)
       }
-      if (log.isDebugEnabled) {
-        log.debug("[{}] Deleting snapshots matching criteria [{}]", currentState, deletionCriteria)
-      }
-      deleteSnapshots(deletionCriteria)
     }
   }
